requirements.txt
dateparser==0.5.1
GitPython==2.0.6
jupyter==1.0.0
jupyter-client==4.3.0
jupyter-console==4.1.1
jupyter-core==4.1.0
matplotlib==1.5.3
nbstripout==0.2.9
networkx==1.11
numpy==1.11.0
requests==2.9.1
scikit-learn==0.17.1
scipy==0.17.1
XlsxWriter==0.9.6
wordcloud
tldextract
